django_backend/requirements.txt
django>=4.2,<5.0
djangorestframework>=3.14,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-filter>=23.5,<24.0
django-cors-headers>=4.3,<5.0
python-dotenv>=1.0,<2.0
gunicorn>=21.2,<23.0
whitenoise>=6.6,<7.0
